+++ internal/conns/awsclient.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/hashicorp/terraform-provider-aws/internal/intf"
+	"github.com/hashicorp/terraform-provider-aws/internal/experimental/intf"
 )
 
 // PartitionHostname returns a hostname with the provider domain suffix for the partition
